pyproject.toml
[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[project]
name = "reviewbuddy"
version = "0.1.0"
description = "AI-powered academic paper review feedback before submission"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    { name = "Katarína Kačmarová" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "dspy>=3.1.2",
    "openai>=1.101.0",
    "litellm>=1.77.1",
    "python-dotenv>=1.1.0",
    "pageindex-local",
    "PyMuPDF>=1.26.4",
    "pymupdf4llm>=0.0.24",
    "pdfplumber>=0.11.9",
    "sentence-transformers>=5.3.0",
    "attachments>=0.25.1",
    "openreview-py>=2.0.1",
    "anthropic>=0.94.0",
]

[project.optional-dependencies]
experiments = [
    "torch>=2.0.0",
    "bert-score>=0.3.13",
    "evaluate>=0.4.6",
    "rouge>=1.0.1",
    "absl-py>=1.0.0",
    "nltk>=3.8.0",
    "rouge-score>=0.1.2",
    "scipy>=1.17.0",
    "scikit-learn>=1.8.0",
    "numpy>=2.4.2",
    "transformers>=5.0.0",
]

[tool.uv.build-backend]
module-root = ""
source-include = ["reviewbuddy/rubric_iclr.txt", "experiments/"]
